small update
added in navmeshing and fixed the player selection for the enemy ai. it theoretically should be able to pathfind now but haven't tested in depth yet

also adjusted the first enemies boxes, made them invisible so the player isn't staring at a weird grey box when they exit. still works as normal.

going to try adjusting the first enemy soon to maybe add animations/give a genuine game-over on touch